Moving screens not currently being used into their own folders.
