+++ __tests__/Screens/Account_AddBankCard.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { describe, expect, test } from "@jest/globals";
 import { render, fireEvent } from "@testing-library/react-native";
 
-import AccountAddBankCard from "../../src/screens/Account_AddBankCard";
+import AccountAddBankCard from "../../src/screens/Currently Not in User Flow/Account_AddBankCard";
 
 it("renders", () => {
   render(<AccountAddBankCard key={"Test"} name={"Test"} />);
