+++ __tests__/Screens/Account_AddCryptoPaymentCard.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { describe, expect, test } from "@jest/globals";
 import { render, fireEvent } from "@testing-library/react-native";
 
-import AccountAddCrptoPaymentCard from "../../src/screens/Account_AddCryptoPaymentCard";
+import AccountAddCrptoPaymentCard from "../../src/screens/Currently Not in User Flow/Account_AddCryptoPaymentCard";
 
 it("renders", () => {
   render(<AccountAddCrptoPaymentCard key={"Test"} name={"Test"} />);
