+++ src/screens/Currently Not in User Flow/PurchaseSell.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { useCallback, useEffect, useState } from "react";
 import { RouteProp } from "@react-navigation/native";
-import { WithMainBg } from "../Components/WithMainBg";
+import { WithMainBg } from "../../Components/WithMainBg";
 import {
   View,
   Text,
@@ -11,9 +11,9 @@ import {
   StyleSheet,
 } from "react-native";
 import Svg, { Path, Defs, G, Pattern, Use, Image } from "react-native-svg";
-import BigBlueButton from "../Components/BigBlueButton";
-import { CashLogoSVG } from "../icons/crypto_icons";
-import { OneFiLogoSVG } from "../icons/crypto_icons";
+import BigBlueButton from "../../Components/BigBlueButton";
+import { CashLogoSVG } from "../../icons/crypto_icons";
+import { OneFiLogoSVG } from "../../icons/crypto_icons";
 
 const backgroundImage = require("../../assets/backgroundSquare.png");
 interface OwnProps {}
